将单图层从std::vector<std::unique_ptr<DrawBase>>更改为单独的图层类，为外部生命周期的渲染包提供接口；文本回看功能推进中

+++ EbbGlow/source/EbbGlow/UI/Animation/KeyFrameAnimation.cpp
@@ -61,7 +61,7 @@ namespace ebbglow::ui
 						if (!activeCom.loop)
 						{
 							inactiveCom.stop = true;
-							inactiveCom.activeFrame = activeCom.keyFrames.size() - 1;
+							inactiveCom.activeFrame = static_cast<int>(activeCom.keyFrames.size()) - 1;
 						}
 						else
 						{
